Training_data/requirements.txt
opencv-contrib-python
numpy
matplotlib
tensorflow
keras
torch
python-socketio
Flask
eventlet
